+++ hank-server/src/main/java/com/liveramp/hank/storage/cueball/Cueball.java
@@ -187,7 +187,7 @@ public class Cueball extends IncrementalStorageEngine implements StorageEngine {
         hashIndexBits,
         getCompressionCodec(),
         configurator.getCacheNumBytesCapacity(),
-        configurator.getCacheNumItemsCapacity());
+        (int)configurator.getCacheNumItemsCapacity());
   }
 
   private CueballCompressionCodec getCompressionCodec() throws IOException {
